Add get_outputs function
This allows for using outputs from one stack as inputs for another. The vpc-peering example has been updated to use it. Also add connection options to the example as it seems to be a common use case. Required some fixes to how the test class (and get_outputs) runs Terraform and Pretf because it was going too many levels deep and the use of os.chdir was causing issues.